feat: Modify some logics of Modal component
- 상위 컴포넌트인 캘린더와의 병합을 고려하여 모달 컴포넌트 구조 수정
