+++ public/components/detail-modal/js/index.js
@@ -1,1 +1,1 @@
-export { default as Modal } from './Modal.js';
+export { default as Modal } from '../Modal.js';
